requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
boto3==1.40.54
botocore==1.40.54
gunicorn==23.0.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
Werkzeug==3.0.3
